Be consistent about the return type from std::min for the microsoft compiler (https://trac.osgeo.org/gdal/ticket/6704).
vrtderivedrasterband.cpp(471): error C2782: 'const _Ty &std::min(const _Ty &,const _Ty &) noexcept(<expr>)': template parameter '_Ty' is ambiguous


git-svn-id: https://svn.osgeo.org/gdal/trunk@36006 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/vrt/vrtderivedrasterband.cpp
@@ -468,10 +468,10 @@ static bool LoadPythonAPI()
     EnumProcessModules(hProcess, ahModules, sizeof(ahModules),
                         &nSizeNeeded);
 
-    const int nModules =
+    const size_t nModules =
         std::min(size_t(100),
                  static_cast<size_t>(nSizeNeeded) / sizeof(HMODULE));
-    for(int i=0;i<nModules;i++)
+    for( size_t i = 0; i < nModules; i++ )
     {
         if( GetProcAddress(ahModules[i], "Py_SetProgramName") )
         {
